+++ src/refactoring/App.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { CartPage } from './components/CartPage.tsx';
+import { CartPage } from './pages/CartPage.tsx';
 import { AdminPage } from './components/AdminPage.tsx';
 import { Coupon, Product } from '../types.ts';
 import { useCoupons, useProducts } from './hooks';
